# Conflicts:
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/slu/com/pandora/activity/KitchenActivity.java
#	app/src/main/java/slu/com/pandora/activity/Login.java
#	app/src/main/java/slu/com/pandora/activity/OrderActivity.java
#	app/src/main/java/slu/com/pandora/adapter/CurrentOrderAdapter.java
#	app/src/main/java/slu/com/pandora/fragment/CurrentOrdersFragment.java
#	app/src/main/java/slu/com/pandora/fragment/QueueOrdersFragment.java
#	app/src/main/java/slu/com/pandora/holder/CurrentOrderHolder.java
#	app/src/main/java/slu/com/pandora/model/Order.java
#	app/src/main/java/slu/com/pandora/model/OrderResponse.java
#	app/src/main/java/slu/com/pandora/rest/ApiClient.java
#	app/src/main/java/slu/com/pandora/rest/ApiInterface.java
#	app/src/main/res/layout-land/current_list.xml
#	app/src/main/res/layout-land/order_layout.xml
#	app/src/main/res/layout-land/sample_listview.xml
#	app/src/main/res/layout/confirm_order_list_view_row.xml
#	app/src/main/res/layout/current_header.xml
#	app/src/main/res/layout/popup_send_order.xml
#	app/src/main/res/layout/queue_finished_list.xml
#	app/src/main/res/menu/appbar_order.xml
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/slu/com/pandora/model/Order.java
@@ -14,7 +14,7 @@ public class Order {
     private Integer id;
     @SerializedName("total")
     @Expose
-    private Double total;
+    private Integer total;
     @SerializedName("tablenum")
     @Expose
     private Integer tablenum;
@@ -35,11 +35,11 @@ public class Order {
         this.id = id;
     }
 
-    public Double getTotal() {
+    public Integer getTotal() {
         return total;
     }
 
-    public void setTotal(Double total) {
+    public void setTotal(Integer total) {
         this.total = total;
     }
